requirements_dev.txt
numpy==1.22.3
pyserial==3.5
PySide6==6.3.0
PySide6-Addons==6.3.0
PySide6-Essentials==6.3.0
shiboken6==6.3.0
